+++ Source/Frontend/UI/Input/Input.cs
@@ -69,22 +69,6 @@
             GamePad.Cleanup();
         }
 
-        public enum InputEventType
-        {
-            Press, Release
-        }
-
-        public class InputEvent
-        {
-            public LogicalButton LogicalButton;
-            public InputEventType EventType;
-
-            public override string ToString()
-            {
-                return string.Format("{0}:{1}", EventType.ToString(), LogicalButton.ToString());
-            }
-        }
-
         private readonly WorkingDictionary<string, object> ModifierState = new WorkingDictionary<string, object>();
         private readonly WorkingDictionary<string, bool> LastState = new WorkingDictionary<string, bool>();
         private readonly WorkingDictionary<string, bool> UnpressState = new WorkingDictionary<string, bool>();
